Breakup /hybrid/ctl WH Playbook
The /hybrid/ctl playbook 'update-hybrid' checksout the latest main
branch AND runs hybrid.yml, however it runs hybrid.yml as it was when
the command executed, which is probably a good thing but it's not what
we want.
To checkout the latest changes THEN run hybrid.yml we need to create and
run 2 seperate ansible.Runners, one to update git, the other to run the
playbook normally.