Isolate acme.sh Role to rosetta:&rosetta.systems
Currently any host set to group rosetta can run the acme.sh role, which
is not configured to create a test tls cert, so any non-production hosts
fail a play containing the role. Additionally the nginx role checks for
the files generated by acme.sh before reloading the config.

As a temporary workaround we:
  1) Specify hosts: rosetta:&rosetta.systems to run the acme.sh role
  2) Created dummy cert files on the staging server to satisfy the nginx
     role as it only check that the specified named files exist, and
     does no validation

A permanent and stable solution would configure acme.she role to support
a testing/staging environment