Setup 'stage' branch for Maneuvers development
Currently development of maneuvers is difficult because testing changes
requires changing the github repository each time we want to run a
playbook, which is obviously undesireable.

To fix this we now have a 'stage' branch to push development changes we
want to run against the staging server, this allows us to push any
changes we want to 'stage', change history, test plays, etc. without
trashing up our production branch. The only thing thats required is
passing in `-e "maneuvers_branch=stage"` when running the playbook.